* Remove duplicated code
