+++ app/src/main/java/com/hisham/ctintegrationsample/searchlist/AbsSearchFragment.kt
@@ -1,0 +1,83 @@
+package com.hisham.ctintegrationsample.searchlist
+
+import android.os.Bundle
+import android.view.*
+import androidx.appcompat.widget.SearchView
+import androidx.lifecycle.Observer
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.NavigationUI
+import androidx.recyclerview.widget.DividerItemDecoration
+import com.hisham.ctintegrationsample.BaseFragment
+import com.hisham.ctintegrationsample.R
+import com.hisham.ctintegrationsample.searchlist.data.SearchListItem
+import com.hisham.ctintegrationsample.searchlist.views.SearchListAdapter
+import kotlinx.android.synthetic.main.search_list_view.*
+
+abstract class AbsSearchFragment : BaseFragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        setHasOptionsMenu(true)
+        return inflater.inflate(R.layout.search_list_view, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        toolbarIcon(0)
+
+        initAdapter()
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        super.onCreateOptionsMenu(menu, inflater)
+        inflater.inflate(R.menu.search, menu)
+        val searchItem = menu.findItem(R.id.searchAction)
+        val searchView = searchItem.actionView as SearchView
+
+        searchView.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
+            override fun onQueryTextSubmit(value: String?): Boolean {
+                return false
+            }
+
+            override fun onQueryTextChange(value: String?): Boolean {
+                return false
+            }
+        })
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return NavigationUI.onNavDestinationSelected(item, findNavController())
+                || super.onOptionsItemSelected(item)
+    }
+
+    private fun initAdapter() {
+        viewModel().apply {
+            // Observe
+            liveData.observe(viewLifecycleOwner, Observer {
+                searchListRecyclerView.apply {
+                    setHasFixedSize(true)
+                    addItemDecoration(
+                        DividerItemDecoration(
+                            context,
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+                    adapter = SearchListAdapter(it).apply {
+                        onItemClickListener { item ->
+                            if (item is SearchListItem.Country) {
+                                update(item)
+                                findNavController().navigateUp()
+                            }
+                        }
+                    }
+                }
+            })
+
+            fetchList()
+        }
+    }
+
+    abstract fun viewModel(): AbsSearchViewModel
+}
