--- app/src/main/java/com/hisham/ctintegrationsample/searchlist/BaseSearchFragment.kt
@@ -1,51 +1,0 @@
-package com.hisham.ctintegrationsample.searchlist
-
-import android.os.Bundle
-import android.view.*
-import androidx.appcompat.widget.SearchView
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import androidx.navigation.ui.NavigationUI
-import com.hisham.ctintegrationsample.R
-
-abstract class BaseSearchFragment : Fragment() {
-
-    abstract fun initAdapter()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        setHasOptionsMenu(true)
-        return inflater.inflate(R.layout.search_list_view, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        initAdapter()
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        inflater.inflate(R.menu.search, menu)
-        val searchItem = menu.findItem(R.id.searchAction)
-        val searchView = searchItem.actionView as SearchView
-
-        searchView.setOnQueryTextListener(object: SearchView.OnQueryTextListener {
-            override fun onQueryTextSubmit(value: String?): Boolean {
-                return false
-            }
-
-            override fun onQueryTextChange(value: String?): Boolean {
-                return false
-            }
-        })
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return NavigationUI.onNavDestinationSelected(item, findNavController())
-                || super.onOptionsItemSelected(item)
-    }
-
-
-}
